+++ src/data/codec/crypto/mod.rs
@@ -1,18 +1,23 @@
 // devela::data::codec::crypto
 //
-#![doc = crate::_tags!(wip)]
 #![doc = crate::_DOC_DATA_CODEC_CRYPTO!()] // public
 #![doc = crate::_doc!(modules: crate::data::codec; crypto)]
 #![doc = crate::_doc!(flat:"data")]
 #![doc = crate::_doc!(hr)]
 //
 
-// mod otp;
-// mod sha1;
+mod digest; // Digest
+mod error; // CryptoError
+// mod otp; // WIP
+// mod sha1; // Sha1 WIP
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::otp::*;
-        // pub use super::sha1::*;
+        pub use super::{
+            digest::*,
+            error::*,
+            // otp::*,
+            // sha1::*,
+        };
     }
 }