+++ src/data/codec/mod.rs
@@ -33,13 +33,13 @@ mod radix; // radix-based encodings (Base32, Base64, Base58…).
 mod types;
 
 // mod compress; // WIP compression algorithms
-// pub mod crypto; // cryptography
+pub mod crypto; // Cryptographic primitives (Digest, Sha1)
 // mod _wip_crc; // WIP
 // pub mod frame; // WIP
-pub mod hash; // hashing algorithms (Fnv, Fx, MD5).
-// mod hex; // WIP Hexadecimal literals and conversions.
-// mod rle; // WIP Run-length encoding and similar techniques.
-mod deser; // WIP structured serialization/deserialization.
+pub mod hash; // hashing algorithms (Fnv, Fx, MD5)
+// mod hex; // WIP Hexadecimal literals and conversions
+// mod rle; // WIP Run-length encoding and similar techniques
+mod deser; // WIP structured serialization/deserialization
 // pub mod schema; // WIP
 
 // #[cfg(feature = "alloc")]
@@ -58,7 +58,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     }
     _pub_mods {
         pub use super::{
-            // crypto::_all::*,
+            crypto::_all::*,
             // frame::_all::*,
             hash::_all::*,
             // schema::_all::*,